cts(compat): MSVC /W4 /WX portability fixes in phaseY8-Y11 shader sources
These newer shader/execution ports were previously compiled only with clang
on macOS; building cts.exe on Windows (MSVC /W4 /WX) promoted two warning
classes to errors. Behavior-preserving:

- texture_utils.cpp: rename local kLevels -> kMipLevels (C4459: shadowed the
  global cts::kLevels in texture_format.h); make depthClearValue double->float
  narrowing explicit via static_cast<float> (C4244).
- atomics/*.spec.cpp (11 files): make const int = t.param<int64_t>(...)
  narrowing explicit via static_cast<int> (C4244).

No logic, scope, or test changes. Smoke-verified atomicAdd:* pass=64/0.

+++ src/webgpu/shader/execution/expression/call/builtin/atomics/atomicAnd.spec.cpp
@@ -15,8 +15,8 @@ TestGroup<AllFeaturesMaxLimitsGpuTest> testGroup = MakeTestGroup<AllFeaturesMaxL
     "Atomically read, and and store value.");
 
 void runAnd(AllFeaturesMaxLimitsGpuTest& t, bool workgroup) {
-    const int workgroupSize = t.param<int64_t>("workgroupSize");
-    const int dispatchSize = t.param<int64_t>("dispatchSize");
+    const int workgroupSize = static_cast<int>(t.param<int64_t>("workgroupSize"));
+    const int dispatchSize = static_cast<int>(t.param<int64_t>("dispatchSize"));
     const ScalarType scalarType = scalarTypeFromParam(t.param<std::string>("scalarType"));
     const MapId& mapId = mapIdFromParam(t.param<std::string>("mapId"));
     const uint32_t numInvocations = static_cast<uint32_t>(workgroup ? workgroupSize : workgroupSize * dispatchSize);
