+++ src/webgpu/shader/execution/expression/call/builtin/atomics/atomicMax.spec.cpp
@@ -17,8 +17,8 @@ TestGroup<AllFeaturesMaxLimitsGpuTest> testGroup = MakeTestGroup<AllFeaturesMaxL
 CTS_TEST(testGroup, "max_storage")
     .params([](ParamsBuilder u) { return basicParams(u); })
     .fn([](AllFeaturesMaxLimitsGpuTest& t) {
-        const int workgroupSize = t.param<int64_t>("workgroupSize");
-        const int dispatchSize = t.param<int64_t>("dispatchSize");
+        const int workgroupSize = static_cast<int>(t.param<int64_t>("workgroupSize"));
+        const int dispatchSize = static_cast<int>(t.param<int64_t>("dispatchSize"));
         const ScalarType scalarType = scalarTypeFromParam(t.param<std::string>("scalarType"));
         std::vector<uint32_t> expected(2, 0);
         expected[0] = static_cast<uint32_t>(workgroupSize * dispatchSize - 1);
@@ -28,8 +28,8 @@ CTS_TEST(testGroup, "max_storage")
 CTS_TEST(testGroup, "max_workgroup")
     .params([](ParamsBuilder u) { return basicParams(u); })
     .fn([](AllFeaturesMaxLimitsGpuTest& t) {
-        const int workgroupSize = t.param<int64_t>("workgroupSize");
-        const int dispatchSize = t.param<int64_t>("dispatchSize");
+        const int workgroupSize = static_cast<int>(t.param<int64_t>("workgroupSize"));
+        const int dispatchSize = static_cast<int>(t.param<int64_t>("dispatchSize"));
         const ScalarType scalarType = scalarTypeFromParam(t.param<std::string>("scalarType"));
         std::vector<uint32_t> expected(static_cast<size_t>(2 * dispatchSize), 0);
         for (int d = 0; d < dispatchSize; ++d) {
