+++ src/webgpu/shader/execution/expression/call/builtin/atomics/atomicStore.spec.cpp
@@ -17,8 +17,8 @@ TestGroup<AllFeaturesMaxLimitsGpuTest> testGroup = MakeTestGroup<AllFeaturesMaxL
 CTS_TEST(testGroup, "store_storage_basic")
     .params([](ParamsBuilder u) { return mapIdParams(u); })
     .fn([](AllFeaturesMaxLimitsGpuTest& t) {
-        const int workgroupSize = t.param<int64_t>("workgroupSize");
-        const int dispatchSize = t.param<int64_t>("dispatchSize");
+        const int workgroupSize = static_cast<int>(t.param<int64_t>("workgroupSize"));
+        const int dispatchSize = static_cast<int>(t.param<int64_t>("dispatchSize"));
         const ScalarType scalarType = scalarTypeFromParam(t.param<std::string>("scalarType"));
         const MapId& mapId = mapIdFromParam(t.param<std::string>("mapId"));
         const uint32_t numInvocations = static_cast<uint32_t>(workgroupSize * dispatchSize);
@@ -41,8 +41,8 @@ CTS_TEST(testGroup, "store_storage_basic")
 CTS_TEST(testGroup, "store_workgroup_basic")
     .params([](ParamsBuilder u) { return mapIdParams(u); })
     .fn([](AllFeaturesMaxLimitsGpuTest& t) {
-        const int workgroupSize = t.param<int64_t>("workgroupSize");
-        const int dispatchSize = t.param<int64_t>("dispatchSize");
+        const int workgroupSize = static_cast<int>(t.param<int64_t>("workgroupSize"));
+        const int dispatchSize = static_cast<int>(t.param<int64_t>("dispatchSize"));
         const ScalarType scalarType = scalarTypeFromParam(t.param<std::string>("scalarType"));
         const MapId& mapId = mapIdFromParam(t.param<std::string>("mapId"));
         const uint32_t numInvocations = static_cast<uint32_t>(workgroupSize);
@@ -66,8 +66,8 @@ CTS_TEST(testGroup, "store_workgroup_basic")
 CTS_TEST(testGroup, "store_storage_advanced")
     .params([](ParamsBuilder u) { return mapIdParams(u); })
     .fn([](AllFeaturesMaxLimitsGpuTest& t) {
-        const int workgroupSize = t.param<int64_t>("workgroupSize");
-        const int dispatchSize = t.param<int64_t>("dispatchSize");
+        const int workgroupSize = static_cast<int>(t.param<int64_t>("workgroupSize"));
+        const int dispatchSize = static_cast<int>(t.param<int64_t>("dispatchSize"));
         const ScalarType scalarType = scalarTypeFromParam(t.param<std::string>("scalarType"));
         const MapId& mapId = mapIdFromParam(t.param<std::string>("mapId"));
         const uint32_t numInvocations = static_cast<uint32_t>(workgroupSize * dispatchSize);
@@ -96,8 +96,8 @@ CTS_TEST(testGroup, "store_storage_advanced")
 CTS_TEST(testGroup, "store_workgroup_advanced")
     .params([](ParamsBuilder u) { return mapIdParams(u); })
     .fn([](AllFeaturesMaxLimitsGpuTest& t) {
-        const int workgroupSize = t.param<int64_t>("workgroupSize");
-        const int dispatchSize = t.param<int64_t>("dispatchSize");
+        const int workgroupSize = static_cast<int>(t.param<int64_t>("workgroupSize"));
+        const int dispatchSize = static_cast<int>(t.param<int64_t>("dispatchSize"));
         const ScalarType scalarType = scalarTypeFromParam(t.param<std::string>("scalarType"));
         const MapId& mapId = mapIdFromParam(t.param<std::string>("mapId"));
         const uint32_t numInvocations = static_cast<uint32_t>(workgroupSize);
